cleanup

+++ src/skygfx/backend.h
@@ -31,7 +31,7 @@ namespace skygfx
 
 		virtual void clear(const std::optional<glm::vec4>& color, const std::optional<float>& depth,
 			const std::optional<uint8_t>& stencil) = 0;
-		virtual void draw(size_t vertex_count, size_t vertex_offset) = 0;
+		virtual void draw(uint32_t vertex_count, uint32_t vertex_offset) = 0;
 		virtual void drawIndexed(uint32_t index_count, uint32_t index_offset) = 0;
 
 		virtual void readPixels(const glm::ivec2& pos, const glm::ivec2& size, TextureHandle* dst_texture) = 0;
