+++ src/skygfx/backend_d3d11.cpp
@@ -636,7 +636,7 @@ void BackendD3D11::clear(const std::optional<glm::vec4>& color, const std::optio
 	}
 }
 
-void BackendD3D11::draw(size_t vertex_count, size_t vertex_offset)
+void BackendD3D11::draw(uint32_t vertex_count, uint32_t vertex_offset)
 {
 	prepareForDrawing();
 	D3D11Context->Draw((UINT)vertex_count, (UINT)vertex_offset);
