+++ src/skygfx/backend_d3d11.h
@@ -32,7 +32,7 @@ namespace skygfx
 
 		void clear(const std::optional<glm::vec4>& color, const std::optional<float>& depth,
 			const std::optional<uint8_t>& stencil) override;
-		void draw(size_t vertex_count, size_t vertex_offset) override;
+		void draw(uint32_t vertex_count, uint32_t vertex_offset) override;
 		void drawIndexed(uint32_t index_count, uint32_t index_offset) override;
 
 		void readPixels(const glm::ivec2& pos, const glm::ivec2& size, TextureHandle* dst_texture) override;
